TinyVec: Remove extend_from_slice_copied_unchecked
The `push_slice_copied` function will probably be enough for 99.999%
of use cases. It's not worthy to maintain an unchecked version of it.

It was primarly used on `TinyString::push_within_capacity`, since we
knew the inner buffer had capacity to push that slice of bytes.
But that function isn't used that often. And even if we wanted to
optimize it, we could do it using `TinyVec::as_mut_ptr`

Example:
```
fn push_repeat<T, const N: usize>(slice: &[T], n: usize, tv: &mut TinyVec<T, N>) {
    let slice_len = slice.len();
    tv.reserve(slice_len * n);

    let mut spare = tv.as_mut_ptr().add(tv.len());
    for _ in 0..n {
        // Manual copy, we've reserved beforehand
        unsafe {
            core::ptr::copy(
                slice.as_ptr(),
                spare,
                slice_len,
            );
            // Move the pointer to the new end
            spare = spare.add(slice_len);
        }
    }
    // Update the length
    unsafe { tv.set_len(tv.len() + (slice_len * n)) }
}
```

In fact, this is probably more "optimized", since we don't update the
length until the end, unlike `push_slice_copied_unchecked` that updates
the length on every call.