Remove TinyVec::from_vec_reuse_buffer
The logical thing when building a TinyVec from a Vec, is for it to
reuse the vec's buffer. There's no need to allocate a new buffer, and
copy the elements and all that stuff.

If we wanted to shrink the buffer, we already have a function for
that: `TinyVec::shrink_to_fit`.